shoppig site update v1.5

+++ src/main/java/com/example/boardproject/repository/WriteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface WriteRepository extends JpaRepository<Board,Long> {
+public interface WriteRepository extends JpaRepository<Board, Integer> {
 }
